# Add caching for submodules and CMake to Windows CI.