--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -171,11 +171,20 @@ jobs:
     steps:
       - name: "Checking out repository"
         uses: actions/checkout@e2f20e631ae6d7dd3b768f56a5d2af784dd54791 # v2.5.0
+      # Restore caches.
+      - name: "Fetching cache (git submodules)"
+        uses: actions/cache/restore@58c146cc91c5b9e778e71775dfe9bf1442ad9a12 # v3.2.3
         with:
-          submodules: false
-      # Update submodules directly using our preferred options. Fetching git
-      # history and source files for large dependencies is particularly slow on
-      # Windows, so we'd rather not rely on the 'checkout' action for this.
+          path: .git/modules
+          key: gitmodules_all_windows_${{ github.sha }}
+          restore-keys: gitmodules_all_windows
+      - name: "Fetching cache (CMake/ccache)"
+        uses: actions/cache/restore@58c146cc91c5b9e778e71775dfe9bf1442ad9a12 # v3.2.3
+        with:
+          path: ${{ github.workspace }}/.ccache
+          key: ccache_all_windows_${{ github.sha }}
+          restore-keys: ccache_all_windows
+      # Fetch deps.
       - name: "Updating git submodules"
         run: git submodule update --init --jobs 8 --depth 1
       - name: "Setting up Python"
@@ -187,12 +196,32 @@ jobs:
           python3 -m venv .venv
           .venv/Scripts/activate.bat
           python3 -m pip install -r runtime/bindings/python/iree/runtime/build_requirements.txt
+      - name: "Installing requirements"
+        run: choco install ccache --yes
       - name: "Configuring MSVC"
         uses: ilammy/msvc-dev-cmd@7315a94840631165970262a99c72cfb48a65d25d # v1.12.0
+      # Finally: build and run tests.
       - name: "Building IREE"
+        env:
+          IREE_USE_LOCAL_CCACHE: 1
+          IREE_READ_REMOTE_CCACHE: 0
+          CCACHE_DIR: ${{ github.workspace }}/.ccache
         run: ./build_tools/cmake/build_all.sh "${BUILD_DIR}"
       - name: "Testing IREE"
         run: ./build_tools/cmake/ctest_all.sh "${BUILD_DIR}"
+      # Write caches (if configured to) after all other steps are finished.
+      - name: "Saving cache (git submodules)"
+        if: needs.setup.outputs.write-caches == 'true'
+        uses: actions/cache/save@58c146cc91c5b9e778e71775dfe9bf1442ad9a12 # v3.2.3
+        with:
+          path: .git/modules
+          key: gitmodules_all_windows_${{ github.sha }}
+      - name: "Saving cache (CMake/ccache)"
+        if: needs.setup.outputs.write-caches == 'true'
+        uses: actions/cache/save@58c146cc91c5b9e778e71775dfe9bf1442ad9a12 # v3.2.3
+        with:
+          path: ${{ github.workspace }}/.ccache
+          key: ccache_all_windows_${{ github.sha }}
 
   build_test_all_bazel:
     needs: setup