--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -207,7 +207,7 @@ jobs:
           IREE_READ_REMOTE_CCACHE: 0
           IREE_WRITE_REMOTE_CCACHE: 0
           IREE_READ_LOCAL_CCACHE: 1
-          IREE_WRITE_LOCAL_CCACHE: ${{ needs.setup.outputs.write-caches }}
+          IREE_WRITE_LOCAL_CCACHE: 1
           CCACHE_DIR: ${{ github.workspace }}/.ccache
           # A full build is around 2.5GB, but uploads/downloads are slow and
           # we have a limit of 10GB per repository