--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -171,7 +171,11 @@ jobs:
     steps:
       - name: "Checking out repository"
         uses: actions/checkout@e2f20e631ae6d7dd3b768f56a5d2af784dd54791 # v2.5.0
-      # Restore caches unconditionally. Cache misses are okay.
+      # Attempt to restore from caches unconditionally.
+      # Note: these will first try to grab a cache entry for this exact commit
+      #       then they will fall back to the latest for any commit.
+      # We're pretty limited on repository cache space (10GB) relative to cache
+      # entry size (~2.5GB), so this usage is pretty optimistic.
       - name: "Fetching cache (git submodules)"
         uses: actions/cache/restore@58c146cc91c5b9e778e71775dfe9bf1442ad9a12 # v3.2.3
         with: