--- a/build_tools/cmake/build_runtime_emscripten.sh
+++ b/build_tools/cmake/build_runtime_emscripten.sh
@@ -54,6 +54,6 @@ echo "Building sample deps"
 echo "------------------"
 "${CMAKE_BIN?}" --build . --target iree-sample-deps -- -k 0
 
-if (( IREE_USE_CCACHE == 1 ));
+if (( IREE_USE_CCACHE == 1 )); then
   ccache --show-stats
 fi