--- a/build_tools/cmake/setup_ccache.sh
+++ b/build_tools/cmake/setup_ccache.sh
@@ -38,6 +38,7 @@ fi
 if (( IREE_READ_REMOTE_CCACHE == 1 || IREE_USE_LOCAL_CCACHE == 1 )); then
   export CMAKE_C_COMPILER_LAUNCHER=$(which ccache)
   export CMAKE_CXX_COMPILER_LAUNCHER=$(which ccache)
+  $(which ccache) --zero-stats
 fi
 
 if (( IREE_READ_REMOTE_CCACHE == 1 )); then