--- a/build_tools/cmake/setup_ccache.sh
+++ b/build_tools/cmake/setup_ccache.sh
@@ -36,9 +36,12 @@ if (( ${IREE_USE_LOCAL_CCACHE} == 1 && ${IREE_READ_REMOTE_CCACHE} == 1)); then
 fi
 
 if (( IREE_READ_REMOTE_CCACHE == 1 || IREE_USE_LOCAL_CCACHE == 1 )); then
+  export IREE_USE_CCACHE=1
   export CMAKE_C_COMPILER_LAUNCHER=$(which ccache)
   export CMAKE_CXX_COMPILER_LAUNCHER=$(which ccache)
   $(which ccache) --zero-stats
+else
+  export IREE_USE_CCACHE=0
 fi
 
 if (( IREE_READ_REMOTE_CCACHE == 1 )); then