@GMNGeoffrey 2023-01-20
> > Given you were having issues with GitHub's cache action compression, you might want to fiddle with https://ccache.dev/manual/4.7.4.html#config_compression
> 
> I tried 5 (see commits), but observed no change in cache size. Could try higher levels or recompression.
> 
> > As a rule of thumb, use level 5 or lower since higher levels may slow down compilations noticeably. Higher levels are however useful when recompressing the cache with command line option -X/--recompress.

Ah I thought GitHub was doing its own compression? I was suggesting that perhaps it would be better to turn off ccache compression entirely, since GitHub is also compressing it