@ScottTodd 2023-01-20
Ah, but we do need to be careful about disk size during the build if not compressing: https://github.com/iree-org/iree/actions/runs/3963413155/jobs/6791208036
```
Local storage:
  Cache size (GB):  2.27 /  5.00 (45.49%)
+ ccache --show-compression
Total data:            2.3 GB (2.3 GB disk blocks)
Compressed data:       2.3 GB (18.4% of original size)
  Original size:      12.3 GB
  Compression ratio: 5.428 x  (81.6% space savings)
```
(note the default max cache size on disk and the original size)